online-quiz-platform/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.32
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
